friends list is "working"
crashes 100% of the time in gingerbread and randomly in jelly bean. ~_~
might be good enough to demo next week but ugh

+++ projects/android/classygames/src/edu/selu/android/classygames/games/Person.java
@@ -5,7 +5,7 @@ public class Person
 {
 
 
-	private int id;
+	private long id;
 	private String name;
 
 
@@ -23,14 +23,14 @@ public class Person
 	}
 
 
-	public Person(final int id, final String name)
+	public Person(final long id, final String name)
 	{
 		this.id = id;
 		this.name = name;
 	}
 
 
-	public int getId()
+	public long getId()
 	{
 		return id;
 	}
